Allow lazily cloning new keys
This switches from `std` to `hashbrown` HashMap as it offers
`.entry_ref()` which is similar to `.entry()` but allows to only clone
the key if it is not already present.

This should make it easier to add &str to a completely owned
Histogram<String> without having to create a new String for every word
seen.

This makes it less nice to work with Copy types like char,
unfortunately. But I think it is worth the trade-off for the easier
interface with Strings or CompactStrings. We could add another Histogram
 focused on Copy types later.